First Logic in Main Activity
Inserting First Logic in Main Activity.

+++ app/src/main/java/com/fwibowo/mronoyuk/Order.kt
@@ -1,3 +1,3 @@
 package com.fwibowo.mronoyuk
 
-data class Order(var From:String,var To:String,var Distance:Float,var Price:Double,var Time:String)
+data class Order(var From:String,var To:String,var Distance:Double,var Price:Double,var Time:String)
