+++ app/src/main/java/com/fwibowo/mronoyuk/Place.kt
@@ -1,3 +1,3 @@
 package com.fwibowo.mronoyuk
 
-data class Place(var Location:Float,var Name:String)
+data class Place(var Location:Double,var Name:String)
